[Rename] 폴더명 수정 apply -> ApplyFormSection
